updated version of the system catalog table enablement notebook
uses the databricks python sdk now
